+++ ChangeLog.md
@@ -7,6 +7,7 @@ This project's release branch is `master`. This log is written from the perspect
 * Add support for aarch64-darwin (apple silicon macs) and aarch64-linux. See [docs/platform-support.md](./docs/platform-support.md) for more information about what's supported and what isn't on each platform.
 * nixpkgs updated to 22.11
 * *Breaking change*: Disable webkit2gtk backend for reflex-dom by default. To enable it, use the useWebkit2Gtk flag in `default.nix`.
+* *Breaking change*: Remove deprecated cabal2nixResult
 * Haskell Libraries Updates
   * reflex-dom 0.6.3.1